chore: useRepeatEverySecond → useRepeatEvery
